server: move the LOQ init code out of server.go
Release note: None
