+++ pkg/server/server.go
@@ -42,8 +42,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/closedts/sidetransport"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/liveness/livenesspb"
-	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/loqrecovery"
-	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/loqrecovery/loqrecoverypb"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/protectedts"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/protectedts/ptprovider"
 	"github.com/cockroachdb/cockroach/pkg/kv/kvserver/protectedts/ptreconcile"
@@ -51,7 +49,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/rpc"
 	"github.com/cockroachdb/cockroach/pkg/rpc/nodedialer"
-	"github.com/cockroachdb/cockroach/pkg/security"
 	"github.com/cockroachdb/cockroach/pkg/server/debug"
 	"github.com/cockroachdb/cockroach/pkg/server/diagnostics"
 	"github.com/cockroachdb/cockroach/pkg/server/serverpb"
@@ -70,7 +67,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/optionalnodeliveness"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire"
 	_ "github.com/cockroachdb/cockroach/pkg/sql/schemachanger/scjob" // register jobs declared outside of pkg/sql
-	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
 	"github.com/cockroachdb/cockroach/pkg/storage/enginepb"
 	"github.com/cockroachdb/cockroach/pkg/ts"
 	"github.com/cockroachdb/cockroach/pkg/ui"
@@ -1554,71 +1550,6 @@ func (s *Server) PreStart(ctx context.Context) error {
 	)
 
 	return maybeImportTS(ctx, s)
-}
-
-func logPendingLossOfQuorumRecoveryEvents(ctx context.Context, stores *kvserver.Stores) {
-	if err := stores.VisitStores(func(s *kvserver.Store) error {
-		// We are not requesting entry deletion here because we need those entries
-		// at the end of startup to populate rangelog and possibly other durable
-		// cluster-replicated destinations.
-		eventCount, err := loqrecovery.RegisterOfflineRecoveryEvents(
-			ctx,
-			s.Engine(),
-			func(ctx context.Context, record loqrecoverypb.ReplicaRecoveryRecord) (bool, error) {
-				event := record.AsStructuredLog()
-				log.StructuredEvent(ctx, &event)
-				return false, nil
-			})
-		if eventCount > 0 {
-			log.Infof(
-				ctx, "registered %d loss of quorum replica recovery events for s%d",
-				eventCount, s.Ident.StoreID)
-		}
-		return err
-	}); err != nil {
-		// We don't want to abort server if we can't record recovery events
-		// as it is the last thing we need if cluster is already unhealthy.
-		log.Errorf(ctx, "failed to record loss of quorum recovery events: %v", err)
-	}
-}
-
-func publishPendingLossOfQuorumRecoveryEvents(
-	ctx context.Context, stores *kvserver.Stores, stopper *stop.Stopper,
-) {
-	_ = stopper.RunAsyncTask(ctx, "publish-loss-of-quorum-events", func(ctx context.Context) {
-		if err := stores.VisitStores(func(s *kvserver.Store) error {
-			recoveryEventsSupported := s.ClusterSettings().Version.IsActive(ctx,
-				clusterversion.UnsafeLossOfQuorumRecoveryRangeLog)
-			_, err := loqrecovery.RegisterOfflineRecoveryEvents(
-				ctx,
-				s.Engine(),
-				func(ctx context.Context, record loqrecoverypb.ReplicaRecoveryRecord) (bool, error) {
-					sqlExec := func(ctx context.Context, stmt string, args ...interface{}) (int, error) {
-						return s.GetStoreConfig().SQLExecutor.ExecEx(ctx, "", nil,
-							sessiondata.InternalExecutorOverride{User: security.RootUserName()}, stmt, args...)
-					}
-					if recoveryEventsSupported {
-						if err := loqrecovery.UpdateRangeLogWithRecovery(ctx, sqlExec, record); err != nil {
-							return false, errors.Wrap(err,
-								"loss of quorum recovery failed to write RangeLog entry")
-						}
-					}
-					// We only bump metrics as the last step when all processing of events
-					// is finished. This is done to ensure that we don't increase metrics
-					// more than once.
-					// Note that if actual deletion of event fails, it is possible to
-					// duplicate rangelog and metrics, but that is very unlikely event
-					// and user should be able to identify those events.
-					s.Metrics().RangeLossOfQuorumRecoveries.Inc(1)
-					return true, nil
-				})
-			return err
-		}); err != nil {
-			// We don't want to abort server if we can't record recovery events
-			// as it is the last thing we need if cluster is already unhealthy.
-			log.Errorf(ctx, "failed to update range log with loss of quorum recovery events: %v", err)
-		}
-	})
 }
 
 // AcceptClients starts listening for incoming SQL clients over the network.
